SpielerNameController in PlayerNameController umbenannt. Falschen Controllerverweis in fxml korrigiert.
