+++ src/main/java/at/fifthwheel/battleship/PlayerNameController.java
@@ -1,16 +1,13 @@
 package at.fifthwheel.battleship;
-import javafx.application.Application;
-import javafx.geometry.Insets;
-import javafx.geometry.Pos;
-import javafx.scene.Scene;
-import javafx.scene.control.Label;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Button;
+import javafx.scene.control.RadioButton;
 import javafx.scene.control.TextField;
-import javafx.stage.Stage;
-import javafx.scene.layout.*;
-import javafx.scene.control.*;
+import javafx.scene.control.ToggleGroup;
 
-
-public class PlayerNameController extends Application {
+public class PlayerNameController {
 
     private SceneManager sceneManager;
 
@@ -18,69 +15,50 @@ public class PlayerNameController extends Application {
         this.sceneManager = sceneManager;
     }
 
-    @Override
-    public void start(Stage primaryStage) {
-       primaryStage.setTitle("Battleship");
+    @FXML
+    private RadioButton singlePlayerButton;
+    @FXML
+    private RadioButton multiPlayerButton;
+    @FXML
+    private TextField player1NameField;
+    @FXML
+    private TextField player2NameField;
+    @FXML
+    private Button startButton;
 
-       VBox root = new VBox(10);
-       root.setAlignment(Pos.CENTER);
-       root.setPadding(new Insets(20));
+    private final ToggleGroup modeGroup = new ToggleGroup();
 
-       Label titleLabel = new Label("Welcome to Battleship");
-       titleLabel.setStyle("-fx-font-size: 20px; -fx-font-weight: bold");
+    @FXML
+    public void initialize() {
+        singlePlayerButton.setToggleGroup(modeGroup);
+        multiPlayerButton.setToggleGroup(modeGroup);
 
-       ToggleGroup modeGroup = new ToggleGroup();
-       RadioButton singlePlayerButton = new RadioButton("Single Player");
-       RadioButton multiPlayerButton = new RadioButton("Multi Player");
-       singlePlayerButton.setToggleGroup(modeGroup);
-       multiPlayerButton.setToggleGroup(modeGroup);
-       singlePlayerButton.setSelected(true);
+        singlePlayerButton.setOnAction(e -> {
+            player2NameField.setDisable(true);
+            player2NameField.clear();
+        });
 
-       TextField player1NameField = new TextField();
-       player1NameField.setPromptText("Player 1");
-       TextField player2NameField = new TextField();
-       player2NameField.setPromptText("Player 2");
-       player2NameField.setDisable(true);
+        multiPlayerButton.setOnAction(e -> {
+            player2NameField.setDisable(false);
+        });
 
-       singlePlayerButton.setOnAction(e -> {
-           player2NameField.setDisable(true);
-           player2NameField.clear();
-       });
-       multiPlayerButton.setOnAction(e -> {
-           player2NameField.setDisable(false);
-       });
-
-       Button startButton = new Button("Start");
-       startButton.setStyle("-fx-font-size: 20px;");
-
-       startButton.setOnAction(e -> {
-          String player1Name = player1NameField.getText().trim();
-          String player2Name = player2NameField.getText().trim();
-          boolean isSinglePlayer = singlePlayerButton.isSelected();
-
-          if (multiPlayerButton.isSelected() && player2Name.isEmpty()) {
-              showAlert("Error!", "Player 2 Name is required");
-          }
-       });
-       root.getChildren().addAll(
-               titleLabel,
-               singlePlayerButton,
-               multiPlayerButton,
-               player1NameField,
-               player2NameField,
-               startButton
-       );
-       Scene scene = new Scene(root, 400, 300);
-       primaryStage.setScene(scene);
-       primaryStage.show();
+        startButton.setOnAction(e -> handleStart());
     }
+
+    private void handleStart() {
+        String player1Name = player1NameField.getText().trim();
+        String player2Name = player2NameField.getText().trim();
+
+        if (multiPlayerButton.isSelected() && player2Name.isEmpty()) {
+            showAlert("Error!", "Player 2 Name is required");
+            return;
+        }
+    }
+
     private void showAlert(String title, String message) {
         Alert alert = new Alert(Alert.AlertType.ERROR);
         alert.setTitle(title);
         alert.setContentText(message);
         alert.showAndWait();
-    }
-    public static void main(String[] args) {
-        launch(args);
     }
 }
